Remove right sidebar, restyle left sidebar
