+++ src/components/Sidebar.css
@@ -1,7 +1,7 @@
 .sidebar {
     width: 250px;
     text-align: left;
-    margin: 10px;
+    margin: 0 10px;
 }
 
 .primary-text {
