+++ src/components/Sidebar.jsx
@@ -2,11 +2,15 @@ import React, {Component} from 'react';
 import './Sidebar.css'
 
 export default class Sidebar extends Component {
+  renderHeaderOrNull() {
+    return this.props.header ? <h3>{this.props.header}</h3> : null;
+  }
+
   render() {
     return (
       <>
         <div className="sidebar">
-          <h3>{this.props.header}</h3>
+          {this.renderHeaderOrNull()}
           <div className={this.props.style}>
             {this.props.children}
           </div>
