+++ src/pages/Home.jsx
@@ -1,11 +1,13 @@
 import React, {Component} from 'react';
+import {generateParagraphs} from "../helpers";
+import * as Content from "../content/StaticContent";
 
 export default class Home extends Component {
   render() {
 
     return (
       <div className="body-content">
-        Nothing here yet...
+        {generateParagraphs(Content.Welcome)}
       </div>
     )
